almost done with sorter

+++ file_sorter/sorter.py
@@ -40,15 +40,20 @@ def del_extra_POST(path):
 
 def recursive_copy(folder_path, output_path):
     contents = os.listdir(folder_path)
-    print(folder_path)
-    print(output_path)
+    #print(folder_path)
+    #print(output_path)
     for content in contents:
             content_path = ap(folder_path, content)
             if os.path.isdir(content_path) and content.find('.zip') == -1:
                 recursive_copy(content_path, output_path)
             elif content.find('.zip') == -1:
-                if folder_path == 'C:\\Users\\riley\\Desktop\\Cows\\Editable Files\\FLIR IMAGES\\Post FLIR': #  Testing
-                    shutil.move(content_path, output_path)
+                    cur_path = content_path.split('\\')
+                    cur_path.append(cur_path[-1].split('//'))
+                    file_name = cur_path[-1][-1]
+                    if os.path.exists(ap(output_path, file_name)):
+                        print('Warning: ' + file_name + ' already exists')
+                    else:
+                        shutil.move(content_path, output_path)
             else:
                 print('Error, cannot handle: \n\t' + content_path)
 
@@ -58,25 +63,28 @@ def recursive_copy(folder_path, output_path):
 ################################################################################
 ################################################################################
 
-def setup():
-    mainPath = get_path() + 'cows\\Editable Files\\FLIR IMAGES'
-    outPath = get_path() + 'cows\\file_sorter\\output_files'
-    folders = os.listdir(mainPath)
+def setup_files():
+    mainPath = get_path() + 'Editable Files\\FLIR IMAGES'
+    outPath = get_path() + 'file_sorter\\output_files'
 
-    del_extra_folders(mainPath)
-    del_extra_POST(mainPath)
+    if os.path.exists(mainPath):
+        folders = os.listdir(mainPath)
+        del_extra_folders(mainPath)
+        del_extra_POST(mainPath)
 
-    for folder in folders:
-        del_non_jpgs(ap(mainPath, folder))
+        for folder in folders:
+            del_non_jpgs(ap(mainPath, folder))
 
-    if os.path.isdir(outPath):
-        shutil.rmtree(outPath)
-    os.makedirs(outPath)
-
-    recursive_copy(mainPath, outPath)
-
+        if os.path.isdir(outPath):
+            shutil.rmtree(outPath)
+            print('tree gone')
+        os.makedirs(outPath)
+        recursive_copy(mainPath, outPath)
+        shutil.rmtree(mainPath)
+    else:
+        print('Files Already Setup :)')
 ################################################################################
 ################################################################################
 ################################################################################
 
-setup()
+setup_files()
